+++ tree_test.go
@@ -12,7 +12,6 @@ import (
 	"strconv"
 	"testing"
 
-	"cosmossdk.io/log"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
@@ -54,7 +53,7 @@ func TestVersionedRandomTree(t *testing.T) {
 	d, closeDB := getTestDB()
 	defer closeDB()
 
-	tree := NewMutableTree(d, 100, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 100, false, NewNopLogger())
 	versions := 50
 	keysPerVersion := 30
 
@@ -130,7 +129,7 @@ func TestTreeHash(t *testing.T) {
 	require.Len(t, expectHashes, versions, "must have expected hashes for all versions")
 
 	r := rand.New(rand.NewSource(randSeed))
-	tree := NewMutableTree(dbm.NewMemDB(), 0, false, log.NewNopLogger())
+	tree := NewMutableTree(dbm.NewMemDB(), 0, false, NewNopLogger())
 
 	keys := make([][]byte, 0, versionOps)
 	for i := 0; i < versions; i++ {
@@ -180,7 +179,7 @@ func TestVersionedRandomTreeSmallKeys(t *testing.T) {
 	d, closeDB := getTestDB()
 	defer closeDB()
 
-	tree := NewMutableTree(d, 100, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 100, false, NewNopLogger())
 	singleVersionTree := getTestTree(0)
 	versions := 20
 	keysPerVersion := 50
@@ -228,7 +227,7 @@ func TestVersionedRandomTreeSmallKeysRandomDeletes(t *testing.T) {
 	d, closeDB := getTestDB()
 	defer closeDB()
 
-	tree := NewMutableTree(d, 100, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 100, false, NewNopLogger())
 	singleVersionTree := getTestTree(0)
 	versions := 30
 	keysPerVersion := 50
@@ -319,7 +318,7 @@ func TestVersionedEmptyTree(t *testing.T) {
 	d, closeDB := getTestDB()
 	defer closeDB()
 
-	tree := NewMutableTree(d, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 0, false, NewNopLogger())
 
 	hash, v, err := tree.SaveVersion()
 	require.NoError(err)
@@ -359,7 +358,7 @@ func TestVersionedEmptyTree(t *testing.T) {
 	tree.Set([]byte("k"), []byte("v"))
 
 	// Now reload the tree.
-	tree = NewMutableTree(d, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(d, 0, false, NewNopLogger())
 	tree.Load()
 
 	require.False(tree.VersionExists(1))
@@ -375,7 +374,7 @@ func TestVersionedTree(t *testing.T) {
 	d, closeDB := getTestDB()
 	defer closeDB()
 
-	tree := NewMutableTree(d, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 0, false, NewNopLogger())
 
 	// We start with empty database.
 	require.Equal(0, tree.ndb.size())
@@ -425,7 +424,7 @@ func TestVersionedTree(t *testing.T) {
 
 	// Recreate a new tree and load it, to make sure it works in this
 	// scenario.
-	tree = NewMutableTree(d, 100, false, log.NewNopLogger())
+	tree = NewMutableTree(d, 100, false, NewNopLogger())
 	_, err = tree.Load()
 	require.NoError(err)
 
@@ -478,7 +477,7 @@ func TestVersionedTree(t *testing.T) {
 	require.EqualValues(hash3, hash4)
 	require.NotNil(hash4)
 
-	tree = NewMutableTree(d, 100, false, log.NewNopLogger())
+	tree = NewMutableTree(d, 100, false, NewNopLogger())
 	_, err = tree.Load()
 	require.NoError(err)
 
@@ -593,7 +592,7 @@ func TestVersionedTreeVersionDeletingEfficiency(t *testing.T) {
 	d, closeDB := getTestDB()
 	defer closeDB()
 
-	tree := NewMutableTree(d, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 0, false, NewNopLogger())
 
 	tree.Set([]byte("key0"), []byte("val0"))
 	tree.Set([]byte("key1"), []byte("val0"))
@@ -685,7 +684,7 @@ func TestVersionedTreeSpecialCase(t *testing.T) {
 	d, closeDB := getTestDB()
 	defer closeDB()
 
-	tree := NewMutableTree(d, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 0, false, NewNopLogger())
 
 	tree.Set([]byte("key1"), []byte("val0"))
 	tree.Set([]byte("key2"), []byte("val0"))
@@ -709,7 +708,7 @@ func TestVersionedTreeSpecialCase2(t *testing.T) {
 	require := require.New(t)
 
 	d := dbm.NewMemDB()
-	tree := NewMutableTree(d, 100, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 100, false, NewNopLogger())
 
 	tree.Set([]byte("key1"), []byte("val0"))
 	tree.Set([]byte("key2"), []byte("val0"))
@@ -722,7 +721,7 @@ func TestVersionedTreeSpecialCase2(t *testing.T) {
 	tree.Set([]byte("key2"), []byte("val2"))
 	tree.SaveVersion()
 
-	tree = NewMutableTree(d, 100, false, log.NewNopLogger())
+	tree = NewMutableTree(d, 100, false, NewNopLogger())
 	_, err := tree.Load()
 	require.NoError(err)
 
@@ -766,7 +765,7 @@ func TestVersionedTreeSpecialCase3(t *testing.T) {
 func TestVersionedTreeSaveAndLoad(t *testing.T) {
 	require := require.New(t)
 	d := dbm.NewMemDB()
-	tree := NewMutableTree(d, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 0, false, NewNopLogger())
 
 	// Loading with an empty root is a no-op.
 	tree.Load()
@@ -790,7 +789,7 @@ func TestVersionedTreeSaveAndLoad(t *testing.T) {
 	require.Equal(int64(6), tree.Version())
 
 	// Reload the tree, to test that roots and orphans are properly loaded.
-	ntree := NewMutableTree(d, 0, false, log.NewNopLogger())
+	ntree := NewMutableTree(d, 0, false, NewNopLogger())
 	ntree.Load()
 
 	require.False(ntree.IsEmpty())
@@ -907,7 +906,7 @@ func TestVersionedCheckpointsSpecialCase3(_ *testing.T) {
 }
 
 func TestVersionedCheckpointsSpecialCase4(t *testing.T) {
-	tree := NewMutableTree(dbm.NewMemDB(), 0, false, log.NewNopLogger())
+	tree := NewMutableTree(dbm.NewMemDB(), 0, false, NewNopLogger())
 
 	tree.Set([]byte("U"), []byte("XamDUtiJ"))
 	tree.Set([]byte("A"), []byte("UkZBuYIU"))
@@ -1026,7 +1025,7 @@ func TestVersionedCheckpointsSpecialCase7(_ *testing.T) {
 
 func TestVersionedTreeEfficiency(t *testing.T) {
 	require := require.New(t)
-	tree := NewMutableTree(dbm.NewMemDB(), 0, false, log.NewNopLogger())
+	tree := NewMutableTree(dbm.NewMemDB(), 0, false, NewNopLogger())
 	versions := 20
 	keysPerVersion := 100
 	keysAddedPerVersion := map[int]int{}
@@ -1154,7 +1153,7 @@ func TestOrphans(t *testing.T) {
 	// Then randomly delete versions other than the first and last until only those two remain
 	// Any remaining orphan nodes should either have fromVersion == firstVersion || toVersion == lastVersion
 	require := require.New(t)
-	tree := NewMutableTree(dbm.NewMemDB(), 100, false, log.NewNopLogger())
+	tree := NewMutableTree(dbm.NewMemDB(), 100, false, NewNopLogger())
 
 	NUMVERSIONS := 100
 	NUMUPDATES := 100
@@ -1305,7 +1304,7 @@ func TestOverwrite(t *testing.T) {
 	require := require.New(t)
 
 	mdb := dbm.NewMemDB()
-	tree := NewMutableTree(mdb, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(mdb, 0, false, NewNopLogger())
 
 	// Set one kv pair and save version 1
 	tree.Set([]byte("key1"), []byte("value1"))
@@ -1318,7 +1317,7 @@ func TestOverwrite(t *testing.T) {
 	require.NoError(err, "SaveVersion should not fail")
 
 	// Reload tree at version 1
-	tree = NewMutableTree(mdb, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(mdb, 0, false, NewNopLogger())
 	_, err = tree.LoadVersion(int64(1))
 	require.NoError(err, "LoadVersion should not fail")
 
@@ -1337,7 +1336,7 @@ func TestOverwriteEmpty(t *testing.T) {
 	require := require.New(t)
 
 	mdb := dbm.NewMemDB()
-	tree := NewMutableTree(mdb, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(mdb, 0, false, NewNopLogger())
 
 	// Save empty version 1
 	_, _, err := tree.SaveVersion()
@@ -1371,7 +1370,7 @@ func TestLoadVersionForOverwriting(t *testing.T) {
 	require := require.New(t)
 
 	mdb := dbm.NewMemDB()
-	tree := NewMutableTree(mdb, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(mdb, 0, false, NewNopLogger())
 
 	maxLength := 100
 	for count := 1; count <= maxLength; count++ {
@@ -1382,10 +1381,10 @@ func TestLoadVersionForOverwriting(t *testing.T) {
 		require.NoError(err, "SaveVersion should not fail")
 	}
 
-	tree = NewMutableTree(mdb, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(mdb, 0, false, NewNopLogger())
 	require.Error(tree.LoadVersionForOverwriting(int64(maxLength * 2)))
 
-	tree = NewMutableTree(mdb, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(mdb, 0, false, NewNopLogger())
 	err := tree.LoadVersionForOverwriting(int64(maxLength / 2))
 	require.NoError(err, "LoadVersion should not fail")
 
@@ -1408,7 +1407,7 @@ func TestLoadVersionForOverwriting(t *testing.T) {
 	require.NoError(err, "SaveVersion should not fail, overwrite was allowed")
 
 	// Reload tree at version 50, the latest tree version is 52
-	tree = NewMutableTree(mdb, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(mdb, 0, false, NewNopLogger())
 	_, err = tree.LoadVersion(int64(maxLength / 2))
 	require.NoError(err, "LoadVersion should not fail")
 
@@ -1444,7 +1443,7 @@ func BenchmarkTreeLoadAndDelete(b *testing.B) {
 	defer d.Close()
 	defer os.RemoveAll("./bench.db")
 
-	tree := NewMutableTree(d, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(d, 0, false, NewNopLogger())
 	for v := 1; v < numVersions; v++ {
 		for i := 0; i < numKeysPerVersion; i++ {
 			tree.Set([]byte(iavlrand.RandStr(16)), iavlrand.RandBytes(32))
@@ -1455,7 +1454,7 @@ func BenchmarkTreeLoadAndDelete(b *testing.B) {
 	b.Run("LoadAndDelete", func(b *testing.B) {
 		for n := 0; n < b.N; n++ {
 			b.StopTimer()
-			tree = NewMutableTree(d, 0, false, log.NewNopLogger())
+			tree = NewMutableTree(d, 0, false, NewNopLogger())
 			require.NoError(b, err)
 			runtime.GC()
 			b.StartTimer()
@@ -1479,7 +1478,7 @@ func BenchmarkTreeLoadAndDelete(b *testing.B) {
 func TestLoadVersionForOverwritingCase2(t *testing.T) {
 	require := require.New(t)
 
-	tree := NewMutableTree(dbm.NewMemDB(), 0, false, log.NewNopLogger())
+	tree := NewMutableTree(dbm.NewMemDB(), 0, false, NewNopLogger())
 
 	for i := byte(0); i < 20; i++ {
 		tree.Set([]byte{i}, []byte{i})
@@ -1541,7 +1540,7 @@ func TestLoadVersionForOverwritingCase2(t *testing.T) {
 func TestLoadVersionForOverwritingCase3(t *testing.T) {
 	require := require.New(t)
 
-	tree := NewMutableTree(dbm.NewMemDB(), 0, false, log.NewNopLogger())
+	tree := NewMutableTree(dbm.NewMemDB(), 0, false, NewNopLogger())
 
 	for i := byte(0); i < 20; i++ {
 		tree.Set([]byte{i}, []byte{i})
@@ -1671,7 +1670,7 @@ func Benchmark_GetWithIndex(b *testing.B) {
 
 	const numKeyVals = 100000
 
-	t := NewMutableTree(db, numKeyVals, false, log.NewNopLogger())
+	t := NewMutableTree(db, numKeyVals, false, NewNopLogger())
 
 	keys := make([][]byte, 0, numKeyVals)
 
@@ -1722,7 +1721,7 @@ func Benchmark_GetByIndex(b *testing.B) {
 
 	const numKeyVals = 100000
 
-	t := NewMutableTree(db, numKeyVals, false, log.NewNopLogger())
+	t := NewMutableTree(db, numKeyVals, false, NewNopLogger())
 
 	for i := 0; i < numKeyVals; i++ {
 		key := iavlrand.RandBytes(10)
@@ -1796,7 +1795,7 @@ func TestNodeCacheStatisic(t *testing.T) {
 			stat := &Statistics{}
 			db, err := dbm.NewDB("test", "memdb", "")
 			require.NoError(t, err)
-			mt := NewMutableTree(db, tc.cacheSize, false, log.NewNopLogger(), StatOption(stat))
+			mt := NewMutableTree(db, tc.cacheSize, false, NewNopLogger(), StatOption(stat))
 
 			for i := 0; i < numKeyVals; i++ {
 				key := []byte(strconv.Itoa(i))
@@ -1827,7 +1826,7 @@ func TestEmptyVersionDelete(t *testing.T) {
 	require.NoError(t, err)
 	defer db.Close()
 
-	tree := NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(db, 0, false, NewNopLogger())
 
 	_, err = tree.Set([]byte("key1"), []byte("value1"))
 	require.NoError(t, err)
@@ -1841,7 +1840,7 @@ func TestEmptyVersionDelete(t *testing.T) {
 	require.NoError(t, tree.DeleteVersionsTo(5))
 
 	// Load the tree from disk.
-	tree = NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(db, 0, false, NewNopLogger())
 	v, err := tree.Load()
 	require.NoError(t, err)
 	require.Equal(t, int64(toVersion), v)
@@ -1858,7 +1857,7 @@ func TestReferenceRoot(t *testing.T) {
 	require.NoError(t, err)
 	defer db.Close()
 
-	tree := NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(db, 0, false, NewNopLogger())
 
 	_, err = tree.Set([]byte("key1"), []byte("value1"))
 	require.NoError(t, err)
@@ -1877,7 +1876,7 @@ func TestReferenceRoot(t *testing.T) {
 	require.NoError(t, err)
 
 	// load the tree from disk
-	tree = NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(db, 0, false, NewNopLogger())
 	_, err = tree.Load()
 	require.NoError(t, err)
 	require.Equal(t, int64(2), tree.Version())
@@ -1888,7 +1887,7 @@ func TestReferenceRoot(t *testing.T) {
 	// test the reference root when pruning
 	db, err = dbm.NewDB("test", "memdb", "")
 	require.NoError(t, err)
-	tree = NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(db, 0, false, NewNopLogger())
 
 	_, err = tree.Set([]byte("key1"), []byte("value1"))
 	require.NoError(t, err)
@@ -1904,7 +1903,7 @@ func TestReferenceRoot(t *testing.T) {
 	require.NoError(t, err)
 
 	// load the tree from disk
-	tree = NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(db, 0, false, NewNopLogger())
 	_, err = tree.Load()
 	require.NoError(t, err)
 
@@ -1914,7 +1913,7 @@ func TestReferenceRoot(t *testing.T) {
 	require.NoError(t, err)
 
 	// load the tree from disk to check if the reference root is loaded correctly
-	tree = NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree = NewMutableTree(db, 0, false, NewNopLogger())
 	_, err = tree.Load()
 	require.NoError(t, err)
 	_, err = tree.Set([]byte("key1"), []byte("value2"))
@@ -1927,7 +1926,7 @@ func TestWorkingHashWithInitialVersion(t *testing.T) {
 	defer db.Close()
 
 	initialVersion := int64(100)
-	tree := NewMutableTree(db, 0, false, log.NewNopLogger())
+	tree := NewMutableTree(db, 0, false, NewNopLogger())
 	tree.SetInitialVersion(uint64(initialVersion))
 
 	v := tree.WorkingVersion()
@@ -1945,7 +1944,7 @@ func TestWorkingHashWithInitialVersion(t *testing.T) {
 	require.NoError(t, err)
 
 	// without WorkingHash
-	tree = NewMutableTree(db, 0, false, log.NewNopLogger(), InitialVersionOption(uint64(initialVersion)))
+	tree = NewMutableTree(db, 0, false, NewNopLogger(), InitialVersionOption(uint64(initialVersion)))
 
 	_, err = tree.Set([]byte("key1"), []byte("value1"))
 	require.NoError(t, err)